:hammer: Big reworking of the code, speeds things up
The main change here is that query_one as a method of getting an individual
item by its ID doesn't scale to this sort of level. So here I switch to my
own index of widgets. This maintains the "this is a bad idea" approach of
having one widget per "pixel", while also making the app usable.

In doing this I've also simplified the whole layout of the code.