:sparkles: Add slow and fast versions of the remove test
This adds a container around the "pixels" of the plot, thus allowing two
forms of remove test: the first removes the container, which then removes
all the children, and it does it fast (as of the version of Textual
available right now). The second remove method is still the same as before,
doing a query on the MandelPoint, and then doing a remove, which results in
a tight loop which takes ages (again in the version of Textual available as
of the time of writing).

This will be helpful in testing a change or two we wish to make to Textual
that will vastly speed up the call to remove against a query.